import cleanup, minor changes, moved getAuthorityType
