+++ java/solr_plugins/RangedSearches/src/de/uni_tuebingen/ub/ixTheo/bibleRangeSearch/BibleRangeParser.java
@@ -3,19 +3,18 @@ package de.uni_tuebingen.ub.ixTheo.bibleRangeSearch;
 
 import java.util.Set;
 import java.util.TreeSet;
+
 import org.apache.lucene.search.Query;
 import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.search.QParser;
 import org.apache.solr.search.SyntaxError;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 
 public class BibleRangeParser extends QParser {
     private final static String QUERY_SEPARATOR = " ";
     private final static String DB_FIELD_SEPARATOR = ",";
-    private static Logger logger = LoggerFactory.getLogger(BibleRangeParser.class);
+    //private static Logger logger = LoggerFactory.getLogger(BibleRangeParser.class);
 
     /**
      * Constructor for the QParser
