+++ java/solr_plugins/RangedSearches/src/de/uni_tuebingen/ub/ixTheo/canonesDateRangeSearch/CanonesDateRangeParser.java
@@ -1,19 +1,15 @@
 package de.uni_tuebingen.ub.ixTheo.canonesDateRangeSearch;
 
 
-import java.util.Set;
-import java.util.TreeSet;
 import org.apache.lucene.search.Query;
 import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.search.QParser;
 import org.apache.solr.search.SyntaxError;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 
 public class CanonesDateRangeParser extends QParser {
-    private static Logger logger = LoggerFactory.getLogger(CanonesDateRangeParser.class);
+    //private static Logger logger = LoggerFactory.getLogger(CanonesDateRangeParser.class);
 
     /**
      * @param qstr        The part of the query string specific to this parser
