+++ java/solr_plugins/RangedSearches/src/de/uni_tuebingen/ub/ixTheo/timeAspectDateRangeSearch/TimeAspectDateRange.java
@@ -1,8 +1,6 @@
 package de.uni_tuebingen.ub.ixTheo.timeAspectDateRangeSearch;
 
 
-import java.util.ArrayList;
-import java.util.List;
 import de.uni_tuebingen.ub.ixTheo.rangeSearch.Range;
