+++ java/solr_plugins/RangedSearches/src/de/uni_tuebingen/ub/ixTheo/timeAspectDateRangeSearch/TimeAspectDateRangeWeight.java
@@ -1,25 +1,30 @@
 package de.uni_tuebingen.ub.ixTheo.timeAspectDateRangeSearch;
 
 
-import com.carrotsearch.hppc.IntFloatHashMap;
-import com.carrotsearch.hppc.IntFloatMap;
+import java.io.IOException;
+
 import org.apache.lucene.document.Document;
 import org.apache.lucene.index.LeafReader;
 import org.apache.lucene.index.LeafReaderContext;
-import org.apache.lucene.search.*;
+import org.apache.lucene.search.ConstantScoreWeight;
+import org.apache.lucene.search.DocIdSetIterator;
+import org.apache.lucene.search.Scorer;
+import org.apache.lucene.search.TwoPhaseIterator;
+import org.apache.lucene.search.Weight;
 import org.apache.lucene.util.BitSet;
 import org.apache.lucene.util.BitSetIterator;
 import org.apache.lucene.util.FixedBitSet;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import java.io.IOException;
+
+import com.carrotsearch.hppc.IntFloatHashMap;
+import com.carrotsearch.hppc.IntFloatMap;
+
 import de.uni_tuebingen.ub.ixTheo.rangeSearch.Range;
 import de.uni_tuebingen.ub.ixTheo.rangeSearch.RangeScorer;
 
 
 public class TimeAspectDateRangeWeight extends ConstantScoreWeight {
     private final static float NOT_RELEVANT = Float.NEGATIVE_INFINITY;
-    private final static Logger logger = LoggerFactory.getLogger(TimeAspectDateRangeWeight.class);
+    //private final static Logger logger = LoggerFactory.getLogger(TimeAspectDateRangeWeight.class);
     private final static String FIELD = "time_aspect_ranges";
     private final IntFloatMap scoring = new IntFloatHashMap();
     private final TimeAspectDateRange[] ranges;
