+++ java/solrmarc_mixin/src/de/unituebingen/ub/ubtools/solrmarcMixin/FidSystematikMixin.java
@@ -3,10 +3,8 @@ package de.unituebingen.ub.ubtools.solrmarcMixin;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Set;
-import java.util.logging.Logger;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import javax.xml.crypto.Data;
 import org.marc4j.marc.DataField;
 import org.marc4j.marc.Record;
 import org.marc4j.marc.Subfield;
