+++ java/solrmarc_mixin/src/de/unituebingen/ub/ubtools/solrmarcMixin/IxTheoPublisher.java
@@ -6,7 +6,6 @@ import org.marc4j.marc.Record;
 import org.marc4j.marc.Subfield;
 import org.marc4j.marc.VariableField;
 import org.solrmarc.index.SolrIndexerMixin;
-import de.unituebingen.ub.ubtools.solrmarcMixin.*;
 
 public class IxTheoPublisher extends SolrIndexerMixin {
     private final static Map<String, String> replacements = new LinkedHashMap<>(128);
